template/requirements.txt
asgiref==3.7.1
Django==3.0.6
djangorestframework==3.14.0
django-cors-headers==3.3.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.2
